Update expense.js
delete expense with userid

+++ controller/expense.js
@@ -1,5 +1,7 @@
 const express = require('express');
 const Expense = require('../models/expense');
+const { Where } = require('sequelize/lib/utils');
+const User = require('../models/user');
 const router = express.Router();
 const addExpense = async (req, res) => {
     const { expenseamount, description, category } = req.body;
@@ -9,7 +11,7 @@ const addExpense = async (req, res) => {
     // }
 
     try {
-        const expense = await Expense.create({ expenseamount, description, category });
+        const expense = await Expense.create({ expenseamount, description, category, userId: req.user.id });
         return res.status(201).json({ expense, success: true });
     } catch (err) {
         console.log("error creating expense:",err)
@@ -31,8 +33,13 @@ const deleteExpense = (req, res) => {
     if(expenseid == undefined || expenseid.length === 0) {
       return res.status(400).json({success:false,})
     }
-    Expense.destroy({ where: { id:expenseid } })
-        .then(() => {
+    Expense.destroy({ where: { id:expenseid ,userId:req.user.id} })
+        .then((noofrows) => {
+          if(noofrows===0)
+            {
+              return res.status(404).json({success:false,message:'expense doesnt belong to user'})
+            }
+           
           return res.status(200).json({success:true, message: "Expense deleted successfully" });
         })
         .catch((err) => {
